% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{orderit-package}
\alias{orderit}
\alias{orderit-package}
\title{orderit: KopfLab Internal Package for Ordering}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Sebastian Kopf \email{sebastian.kopf@colorado.edu} (\href{https://orcid.org/0000-0002-2044-0201}{ORCID})

}
\keyword{internal}
