% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_run.R
\name{start_app}
\alias{start_app}
\title{Run the user interface}
\usage{
start_app(
  data_sheet_id,
  data_folder_id,
  gs_key_file,
  user_id,
  user_is_admin,
  user_groups,
  user_name,
  timezone,
  options = list(),
  uiPattern = "/",
  enableBookmarking = "url"
)
}
\arguments{
\item{data_sheet_id}{google data sheet id}

\item{data_folder_id}{google drive folder id}

\item{gs_key_file}{in json format}

\item{timezone}{the timezone to user for datetime calculations}

\item{options}{Named options that should be passed to the \code{runApp} call
(these can be any of the following: "port", "launch.browser", "host", "quiet",
"display.mode" and "test.mode"). You can also specify \code{width} and
\code{height} parameters which provide a hint to the embedding environment
about the ideal height/width for the app.}

\item{uiPattern}{A regular expression that will be applied to each \code{GET}
request to determine whether the \code{ui} should be used to handle the
request. Note that the entire request path must match the regular
expression in order for the match to be considered successful.}

\item{enableBookmarking}{Can be one of \code{"url"}, \code{"server"}, or
\code{"disable"}. The default value, \code{NULL}, will respect the setting from
any previous calls to  \code{\link[shiny:enableBookmarking]{enableBookmarking()}}. See \code{\link[shiny:enableBookmarking]{enableBookmarking()}}
for more information on bookmarking your app.}
}
\description{
Run the user interface
}
