% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_run.R
\name{start_app_server}
\alias{start_app_server}
\title{Run the user interface on a server}
\usage{
start_app_server(
  data_sheet_id,
  data_folder_id,
  gs_key_file,
  user_id = Sys.getenv("SHINYPROXY_USERNAME"),
  log = TRUE,
  debug = TRUE
)
}
\arguments{
\item{data_sheet_id}{google data sheet id}

\item{data_folder_id}{google drive folder id}

\item{gs_key_file}{in json format}
}
\description{
Run the user interface on a server
}
