% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_module_selector_table.R
\name{module_selector_table_server}
\alias{module_selector_table_server}
\title{Selector table server}
\usage{
module_selector_table_server(
  input,
  output,
  session,
  get_data,
  id_column,
  available_columns = list(dplyr::across(dplyr::everything(), identity)),
  visible_columns = c(),
  allow_view_all = FALSE,
  page_lengths = list(c(5, 10, 20, 50, 100, if (allow_view_all) -1), c("5", 10, "20",
    "50", "100", if (allow_view_all) "All")),
  initial_page_length = page_lengths[[1]][1],
  dom = "fltip",
  filter = c("none", "bottom", "top"),
  class = "cell-border stripe hover order-column",
  selection = c("multiple", "single", "none"),
  render_html = c(),
  formatting_calls = list()
)
}
\arguments{
\item{get_data}{reactive context providing the data set}

\item{id_column}{name of the ID column - must have unique values!! make a rownumber or concatenated column if there is no unique identifier, this column does NOT have to be part of available_columns (but can be)}

\item{available_columns}{list of transmute statements to select columns to show}

\item{visible_columns}{integer vector of columns (from what the available_columns selects) that are visible when the table loads (can be changed with the table_columns_button action) - if empty, all columns are visible}

\item{allow_view_all}{whether to allow the "all" option in the page lengths, default is FALSE}

\item{page_lengths}{page length options, first one will be selected}

\item{initial_page_length}{initially selected page length, first entry of the page_lengths by default}

\item{dom}{the available table control elements and their order}

\item{filter}{whether to include column filters - note that this does NOT work for restoring after reload so use with caution if that's a desired feature}

\item{class}{styling of table see class parameter for datatable}

\item{selection}{see parameter for data table (none, single, multiple)}

\item{render_html}{list of columns which should NOT be html escaped (e.g. for links)}

\item{formatting_calls}{list of lists with function e.g. list(list(func = formatCurrency, columns = "x))}
}
\description{
Selector table server
}
